requirements.txt
pygame==2.6.1
pyOpenGL==3.1.7
PyGLM==2.8.2